% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjust_daily_additive.R
\name{adjust_daily_additive}
\alias{adjust_daily_additive}
\title{Adjust Daily Values by Linear Addition}
\usage{
adjust_daily_additive(x, months, mean_change = 0)
}
\arguments{
\item{x}{vector of values (e.g. daily temperature)}

\item{months}{vector of months corresponding to x}

\item{mean_change}{additive factor (scalar or vector of length 12 for monthly)}
}
\description{
Adjust Daily Values by Linear Addition
}

