% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_daily_gamma.R
\name{adjust_daily_gamma}
\alias{adjust_daily_gamma}
\title{Adjust Daily Values by Shifting Gamma Distribution}
\usage{
adjust_daily_gamma(x, months, mean_changes = 1, cv_changes = 1,
  exclude_zeros = TRUE)
}
\arguments{
\item{x}{vector of values (e.g. daily precipitation)}

\item{months}{vector of months corresponding to x}

\item{mean_changes}{adjustment factor for changing mean (scalar or vector of length 12 for monthly)}

\item{cv_changes}{adjustment factor for changing coef. of variation (scalar or vector of length 12 for monthly)}

\item{exclude_zeros}{if TRUE (default), excludes zero values from adjustment}
}
\description{
Adjust Daily Values by Shifting Gamma Distribution
}

