% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjust_daily_to_annual.R
\name{adjust_daily_to_annual}
\alias{adjust_daily_to_annual}
\title{Adjust Daily Values to Match Annual Sums}
\usage{
adjust_daily_to_annual(values_day, years_day, values_yr, years_yr,
  min_ratio = 0.9, max_ratio = 1.1)
}
\arguments{
\item{values_day}{vector of daily values}

\item{years_day}{vector of years corresponding to values_day}

\item{values_yr}{vector of annual sums}

\item{years_yr}{vector of years corresponding to values_yr}

\item{min_ratio}{minimum adjustment ratio}

\item{max_ratio}{maximum adjustment ratio}
}
\description{
Adjust Daily Values to Match Annual Sums
}

