% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{aggregate_maurer_mon}
\alias{aggregate_maurer_mon}
\title{Aggregate monthly timeseries from maurer to annual timesteps}
\usage{
aggregate_maurer_mon(x)
}
\arguments{
\item{x}{data frame containing monthly maurer timeseries}
}
\description{
Aggregate monthly timeseries from maurer to annual timesteps
}
\examples{
aggregate_maurer_mon(get_maurer_mon(42, -71))
}

