% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{ApplyMaxTempDiagnostics}
\alias{ApplyMaxTempDiagnostics}
\title{ApplyMaxTempDiagnostics}
\usage{
ApplyMaxTempDiagnostics(stn_name = NULL, model = NULL, path = NULL)
}
\description{
Function that runs Pearson's residuals graphical diagnostics on Max Temperature model.
}
