% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{ApplyMinTempDiagnostics}
\alias{ApplyMinTempDiagnostics}
\title{ApplyMinTempDiagnostics}
\usage{
ApplyMinTempDiagnostics(stn_name = NULL, model = NULL, path = NULL)
}
\description{
Function that runs Pearson's residuals graphical diagnostics on Min Temperature model.
}
