% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{ApplyPrcpAmtDiagnostics}
\alias{ApplyPrcpAmtDiagnostics}
\title{ApplyPrcpAmtDiagnostics}
\usage{
ApplyPrcpAmtDiagnostics(stn_name = NULL, model = NULL, path = NULL)
}
\description{
Function that runs Pearson's and Anscombe's residuals graphical diagnostics on Precipitation occurrence model.
}
