% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arima_fit_components.R
\name{arima_fit_components}
\alias{arima_fit_components}
\title{Fit ARIMA Models to Wavelet Components}
\usage{
arima_fit_components(components)
}
\arguments{
\item{components}{2-d numeric array of wavelet components generated by wavelet_components()}
}
\description{
Fit ARIMA Models to Wavelet Components
}
\examples{
arima_fit_components(cbind("COMPONENT"=sin(1:50) + rnorm(50) + 10, "NOISE"=rnorm(50)))
}

