% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_models}
\alias{arima_models}
\title{Create arima models of wavelet components}
\usage{
arima_models(components)
}
\arguments{
\item{components}{2-d numeric array of wavelet components generated by wavelet_components}
}
\description{
Create arima models of wavelet components
}
\examples{
arima_models(cbind("COMPONENT"=sin(1:50) + rnorm(50) + 10, "NOISE"=rnorm(50)))
}

