% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as.climate.tibble}
\alias{as.climate.tibble}
\title{Transform object to climate tibble}
\usage{
as.climate.tibble(object, map_cols = list(date = "date", station_id =
  "station_id", tmax = "tmax", tmin = "tmin", prcp = "prcp"))
}
\description{
Transform object to climate tibble
}
