% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as.sim_loc.sf}
\alias{as.sim_loc.sf}
\title{Transform object to simulation locations simple feature (sf)}
\usage{
as.sim_loc.sf(object, crs, map_cols = list(station_id = "station_id",
  longitude = "longitude", latitude = "latitude"))
}
\description{
Transform object to simulation locations simple feature (sf).
}
