% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as.stations.sf}
\alias{as.stations.sf}
\title{Transform object to stations simple feature (sf)}
\usage{
as.stations.sf(object, crs, map_cols = list(station_id = "station_id",
  longitude = "lon_dec", latitude = "lat_dec"))
}
\description{
Transform object to stations simple feature (sf).
}
