% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autorun.R
\name{autorun_from_yaml}
\alias{autorun_from_yaml}
\title{Autorun from yaml}
\usage{
autorun_from_yaml(params_yaml_filename = NULL)
}
\description{
This function allow a autorun from a yaml named autorun_parameters.yml.
}
