% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{calibrate.glmwgen}
\alias{calibrate.glmwgen}
\title{Weather model calibration}
\usage{
calibrate.glmwgen(climate, stations, seasonal_climate = NULL,
  control = glmwgen:::glmwgen_fit_control(), verbose = T)
}
\description{
Fits a weather model from historical data.
}
