% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spatial.R
\name{calibrate_spatial}
\alias{calibrate_spatial}
\title{Weather model calibration}
\usage{
calibrate_spatial(
  climate,
  stations,
  seasonal_climate = NULL,
  control = glmwgen:::glmwgen_fit_control(),
  verbose = F
)
}
\description{
Fits a weather model from historical data.
}
