% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{climate}
\alias{climate}
\title{Climate dataset}
\format{A data.frame with 1,027,540 observations.}
\usage{
climate
}
\description{
This dataset contains daily climate data provided by Scott Steinschneider.
It appears to be GHCND-Daily data for a number of stations, with gaps removed.
Some files end in "MANUAL" so there may have been some manual adjustments.
It will do for now for testing.
}
\section{Dimensions}{


\itemize{
  \item \code{FILE}: name of original file
  \item \code{DATE}: dates as Date objects
}
}

\section{Measures}{


\itemize{
  \item \code{PRCP}: daily precipitation (mm/day)
  \item \code{TMIN}, \code{TMAX}: daily minimum and maximum temperature (degC)
}
}
\examples{
climate
}

