% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data-climate-cities.R
\docType{data}
\name{climate_cities}
\alias{climate_cities}
\title{Climate cities dataset}
\format{A list of length three, each element a data frame with 22,645 rows}
\usage{
climate_cities
}
\description{
This dataset contains daily climate data for three cities in the northeast
from the Maurer dataset. The cities include Boston, Providence, and Worcester.
The dataset is provided as list of data frames. The dataset was generated
using the R script: /scripts/postgres/climate_cities.R
}
\section{Dimensions}{


\itemize{
  \item \code{FILE}: name of original file
  \item \code{DATE}: dates as Date objects
}
}

\section{Measures}{


\itemize{
  \item \code{PRCP}: daily precipitation (mm/day)
  \item \code{TMIN}, \code{TMAX}: daily minimum and maximum temperature (degC)
  \item \code{WIND}: daily mean windspeed (m/s)
}
}
\examples{
climate_cities[['boston']]
}

