% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complete_water_years.R
\name{complete_water_years}
\alias{complete_water_years}
\title{Get Indices of Complete Water Years}
\usage{
complete_water_years(x, start_month = 10)
}
\arguments{
\item{x}{vector of dates}

\item{start_month}{first month of the water year}
}
\description{
Get Indices of Complete Water Years
}
\examples{
dates <- seq.Date(from=as.Date("2000-01-01"), to=as.Date("2004-12-31"), by="1 day")
dates_complete <- dates[complete_water_years(dates, start_month=10)] # 2000-10-01 -> 20014-09-30
}

