% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{DAILY_WEATHER_GENERATOR}
\alias{DAILY_WEATHER_GENERATOR}
\title{Run daily weather generator}
\usage{
DAILY_WEATHER_GENERATOR(k1, num_year_sim, PRCP_FINAL_ANNUAL_SIM, ANNUAL_PRCP,
  WATER_YEAR_A, WATER_YEAR_D, PRCP, TEMP, TMAX, TMIN, DATE_D, MONTH_D, YEAR_D,
  MONTH_DAY_D, wet_spell_change, dry_spell_change, month_list, water_year_start,
  water_year_end)
}
\arguments{
\item{k1}{trial number as integer}

\item{num_year_sim}{number of years for simulation}

\item{PRCP_FINAL_ANNUAL_SIM}{numeric array of simulated annual precipitation}

\item{WATER_YEAR_A}{numeric array of water years for annual observed data}

\item{WATER_YEAR_D}{array of Dates corresponding to Jan 1 of each water year in WATER_YEAR_A}

\item{PRCP}{numeric array of daily precipitation (mm/day?)}

\item{TEMP}{numeric array of daily mean air temperature (degC?)}

\item{TMAX}{numeric array of daily maximum air temperature (degC?)}

\item{TMIN}{numeric array of daily minimum air temperature (degC?)}

\item{DATE_D}{array of Dates corresponding to daily input data (same length as PRCP, TEMP, TMAX, TMIN)}

\item{MONTH_D}{numeric array of months corresponding to DATE_D}

\item{YEAR_D}{numeric array of years corresponding to DATE_D}

\item{MONTH_DAY_D}{numeric matrix of month and day correspond to DATE_D with column names (MONTH_D, DAY_D)}

\item{wet_spell_change}{numeric array of monthly wet spell adjustment factors}

\item{dry_spell_change}{numeric array of monthly dry spell adjustment factors}

\item{month_list}{integer array of month sequence (starts with 10 for water year, 1 for calendar year)}

\item{water_year_start}{initial water year of observed data as integer}

\item{water_year_end}{final water year of observed data as integer}

\item{ANNUALPRCP}{numeric array of observed annual precipitation}
}
\description{
Run daily weather generator
}
\examples{
DAILY_WEATHER_GENERATOR(k1,num_year_sim,PRCP_FINAL_ANNUAL_SIM,ANNUAL_PRCP,WATER_YEAR_A,WATER_YEAR_D,PRCP,TEMP,TMAX,TMIN,DATE_D,MONTH_D,YEAR_D,MONTH_DAY_D,wet_spell_change,dry_spell_change,month_list,water_year_start,water_year_end)
}

