% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mc.R
\name{fit_mc}
\alias{fit_mc}
\title{Fit Markov Chain Transition Matrices to Observed States}
\usage{
fit_mc(x, months)
}
\arguments{
\item{x}{character array of states as ordered factor}

\item{months}{numeric array of months for each daily time step}
}
\description{
Fit Markov Chain Transition Matrices to Observed States
}
\examples{
mats <- fit_mc(x=sample(c('d', 'w', 'e'), size=720, replace=TRUE, prob=c(0.5, 0.3, 0.2)), months=rep(rep(seq(1, 12), each=30), times=2))
}

