% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamwgen-package.R
\docType{package}
\name{gamwgen}
\alias{gamwgen}
\title{gamwgen: generalized aditive model weather generator}
\description{
Weather series generation based on generalized aditive models?
}
\section{Model fitting}{

To be done...
}

