% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gen_month_arma}
\alias{gen_month_arma}
\title{Generate monthly climate timeseries using ARMA model}
\usage{
gen_month_arma(x.reg.yr, x.loc.mon, n.iter, n.year)
}
\arguments{
\item{x.reg.yr}{data.frame with regional average annual climate timeseries (YEAR, PRCP, TMAX, TMIN, TAVG)}

\item{x.loc.mon}{data.frame with annual climate timeseries at selected location (YEAR, MONTH, PRCP, TMAX, TMIN, TAVG)}

\item{n.iter}{number of iterations for arima simulation}

\item{n.year}{number of years for arima simulation}
}
\description{
Generate monthly climate timeseries using ARMA model
}
\examples{
gen_month_arma(x.reg.yr=clim.reg.yr, x.loc.mon=clim.loc.mon, n.iter=50, n.year=nrow(clim.reg.yr))
}

