% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GET_PI}
\alias{GET_PI}
\title{Solve transition matrix}
\usage{
GET_PI(p00, p01, p02, p10, p11, p12, p20, p21, p22)
}
\arguments{
\item{p00}{transition probability from 0 to 0}

\item{p01}{transition probability from 0 to 1}

\item{p02}{transition probability from 0 to 2}

\item{p10}{transition probability from 1 to 0}

\item{p11}{transition probability from 1 to 1}

\item{p12}{transition probability from 1 to 2}

\item{p20}{transition probability from 2 to 0}

\item{p21}{transition probability from 2 to 1}

\item{p22}{transition probability from 2 to 2}
}
\description{
Solve transition matrix
}
\examples{
GET_PI(p00,p01,p02,p10,p11,p12,p20,p21,p22)
}

