% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_package.R
\docType{package}
\name{glmwgen}
\alias{glmwgen}
\alias{glmwgen-package}
\title{glmwgen: general lineal model weather generator}
\description{
Weather series generation based on general linear models?
}
\section{Model fitting}{

To be done...
}

