% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{glmwgen_simulation_control}
\alias{glmwgen_simulation_control}
\title{Simulations control configuration}
\usage{

  glmwgen_simulation_control(seasonal_temps_covariats_getter = get_temperatures_seasonal_covariat,
  seasonal_prcp_covariats_getter = get_rainfall_seasonal_covariat,
  random_fields_method = "gaussian", Rt = NULL, grf_method = NULL,
  multicombine = T)
}
\description{
Provides fine control of different parameters that will be used to create new weather series.
}

