% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{glmwgenControl}
\alias{glmwgenControl}
\title{Weather model fit configuration}
\usage{
glmwgenControl(prcp_occurrence_threshold = 0.1, use_seasonal_covariats = F,
  seasonal_covariats = c("tx", "tn", "prcp"))
}
\description{
Provides fine control of different parameters that will be used to fit a weather model.
}

