% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{KNN}
\alias{KNN}
\title{k-Nearest Neighbors of monthly climate}
\usage{
KNN(cur_sim_PRCP, cur_sim_TEMP, PRCP_TODAY, TEMP_TODAY, PRCP_TOMORROW,
  TEMP_TOMORROW, TMAX_TOMORROW, TMIN_TOMORROW, DATE_TOMORROW, k,
  sd_monthly_PRCP, sd_monthly_TEMP, mean_monthly_PRCP, mean_monthly_TEMP, k1,
  count)
}
\arguments{
\item{cur_sim_PRCP}{value of simulated daily precipitation}

\item{cur_sim_TEMP}{value of simulated daily mean air temperature}

\item{PRCP_TODAY}{amount of precipitation for today}

\item{TEMP_TODAY}{mean air temperature for today}

\item{PRCP_TOMORROW}{amount of precipitation for tomorrow}

\item{TEMP_TOMORROW}{mean air temperature for tomorrow}

\item{TMAX_TOMORROW}{maximum air temperature for tomorrow}

\item{TMIN_TOMORROW}{minimum air temperature for tomorrow}

\item{DATE_TOMORROW}{date of tomorrow}

\item{k}{number of nearest neighbors}

\item{sd_monthly_PRCP}{standard deviation of monthly precip}

\item{sd_monthly_TEMP}{standard deviation of monthly air temperature}

\item{mean_monthly_PRCP}{mean of monthly precip}

\item{mean_monthly_TEMP}{mean of monthly air temperature}

\item{k1}{trial number}

\item{count}{current timestep}
}
\description{
k-Nearest Neighbors of monthly climate
}
\examples{
KNN(cur_sim_PRCP,cur_sim_TEMP,PRCP_TODAY,TEMP_TODAY,PRCP_TOMORROW,TEMP_TOMORROW,TMAX_TOMORROW,TMIN_TOMORROW,DATE_TOMORROW,k,sd_monthly_PRCP,sd_monthly_TEMP,mean_monthly_PRCP,mean_monthly_TEMP,k1,count)
}

