% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{KNN_ANNUAL}
\alias{KNN_ANNUAL}
\title{k-Nearest Neighbors of Annual Sequence}
\usage{
KNN_ANNUAL(sim_annual_prcp, ANNUAL_PRCP, WATER_YEAR_A, kk, k1, y)
}
\arguments{
\item{sim_annual_prcp}{value of simulated annual precipitation}

\item{ANNUAL_PRCP}{numeric array of observed annual precipitation}

\item{WATER_YEAR_A}{integer array of water years corresponding to observed annual precipitation}

\item{kk}{number of nearest neighbors (?)}

\item{k1}{trial number as integer}

\item{y}{integer representing index of year simulation in 1:num_years_sim}
}
\description{
k-Nearest Neighbors of Annual Sequence
}
\examples{
KNN_ANNUAL(sim_annual_prcp,ANNUAL_PRCP,WATER_YEAR_A,kk,k1,y)
}

