% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_daily.R
\name{knn_daily}
\alias{knn_daily}
\title{Select k-Nearest Neighbor for Given Simulation State}
\usage{
knn_daily(x, wday, state, state_prev, prcp_prev, temp_prev, prcp_sd, temp_sd)
}
\arguments{
\item{x}{historical daily climate dataset}

\item{wday}{simulation water-day}

\item{state}{simulated state}

\item{state_prev}{simulated state of previous water-day}

\item{prcp_prev}{simulated precip of previous water-day}

\item{temp_prev}{simulated temp of previous water-day}

\item{prcp_sd}{standard deviation of historical precipitation for current month}

\item{temp_sd}{standard deviation of historical temperature for current month}
}
\description{
Select k-Nearest Neighbor for Given Simulation State
}

