% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_local.R
\name{local_fit_control}
\alias{local_fit_control}
\title{Weather model fit configuration}
\usage{
local_fit_control(
  prcp_occurrence_threshold = 0.1,
  use_external_seasonal_climate = T,
  climate_missing_threshold = 0.2,
  use_covariates = F,
  avbl_cores = 2,
  trim_gam_fit_models = F,
  planar_crs_in_metric_coords = 22185
)
}
\description{
Provides fine control of different parameters that will be used to fit a weather model.
}
