% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maurer_aggregate.R
\name{maurer_aggregate}
\alias{maurer_aggregate}
\title{Aggregate monthly timeseries from maurer to annual timesteps}
\usage{
maurer_aggregate(x)
}
\arguments{
\item{x}{data frame containing monthly maurer timeseries}
}
\description{
Aggregate monthly timeseries from maurer to annual timesteps
}
\examples{
maurer_aggregate(maurer_get_mon(42, -71))

}

