% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maurer_get_month.R
\name{maurer_get_month}
\alias{maurer_get_month}
\title{Retrieve monthly timeseries for given location from maurer dataset}
\usage{
maurer_get_month(lat, lon)
}
\arguments{
\item{lat}{latitude in decimal degrees North}

\item{lon}{longitude in decimal degrees East (negative for West)}
}
\description{
Retrieve monthly timeseries for given location from maurer dataset
}
\examples{
maurer_get_month(42, -71)
}

