% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_adjust_transition.R
\name{mc_adjust_transition}
\alias{mc_adjust_transition}
\title{Adjust Markov Chain Transition Matrix}
\usage{
mc_adjust_transition(transition, dry_spell = 1, wet_spell = 1)
}
\arguments{
\item{transition}{markov transition matrix}

\item{dry_spell}{change factor for dry spell (default=1, no change)}

\item{wet_spell}{change factor for wet spell (default=1, no change)}
}
\description{
Adjust Markov Chain Transition Matrix
}
\examples{
transitions <- mc_fit(x=sample(c('d', 'w', 'e'), size=720, replace=TRUE, prob=c(0.5, 0.3, 0.2)), months=rep(rep(seq(1, 12), each=30), times=2))
mc_adjust_transition(transition[[1]], dry_spell=1, wet_spell=1)
}

