% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/arima.R
\name{mc_arimas}
\alias{mc_arimas}
\title{Run Monte Carlo simulation of arima models}
\usage{
mc_arimas(models, n, n.iter)
}
\arguments{
\item{models}{list of arima model objects (e.g. noise and wavelet components)}

\item{n}{number of simulation timesteps}

\item{n.iter}{number of monte carlo iterations}
}
\description{
Run Monte Carlo simulation of arima models
}
\examples{
library(forecast)
models <- list(COMPONENT=auto.arima(sin(1:50)+rnorm(50)+10, max.p=2, max.q=2, max.P=0, max.Q=0, stationary=TRUE),
               NOISE=auto.arima(rnorm(50), max.p=2, max.q=2, max.P=0, max.Q=0, stationary=TRUE))
mc_arimas(models, n=40, n.iter=100)
}

