% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_assign_states.R
\name{mc_assign_states}
\alias{mc_assign_states}
\title{Assign Markov States from Precipitation and State Thresholds}
\usage{
mc_assign_states(prcp, months, states, thresholds)
}
\arguments{
\item{prcp}{vector of daily precipitation}

\item{months}{vector of months corresponding to daily precipitation}

\item{states}{character list of markov states}

\item{thresholds}{list of monthly transition matrices generated from mc_fit()}
}
\description{
Assign Markov States from Precipitation and State Thresholds
}

