% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mc_fit.R
\name{mc_fit}
\alias{mc_fit}
\title{Fit Markov Chain Transition Matrices to State Sequence}
\usage{
mc_fit(states, months)
}
\arguments{
\item{states}{character array of states as ordered factor}

\item{months}{numeric array of months for each daily time step}
}
\description{
Fit Markov Chain Transition Matrices to State Sequence
}
\examples{
transitions <- mc_fit(x=sample(c('d', 'w', 'e'), size=720, replace=TRUE, prob=c(0.5, 0.3, 0.2)), months=rep(rep(seq(1, 12), each=30), times=2))
}

