% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mc_state_equilibrium.R
\name{mc_state_equilibrium}
\alias{mc_state_equilibrium}
\title{Compute State Equilibrium Vector of Transition Matrix}
\usage{
mc_state_equilibrium(m)
}
\arguments{
\item{m}{markov transition matrix}
}
\description{
Compute State Equilibrium Vector of Transition Matrix
}
\examples{
transitions <- mc_fit(x=sample(c('d', 'w', 'e'), size=720, replace=TRUE, prob=c(0.5, 0.3, 0.2)), months=rep(rep(seq(1, 12), each=30), times=2))
mc_state_equilibrium(transitions[[1]])
}

