% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mc.R
\name{mc_state_threshold}
\alias{mc_state_threshold}
\title{Determine Markov State Thresholds from Precipitation}
\usage{
mc_state_threshold(prcp, months, dry_wet = 0.3, wet_extreme_quantile = 0.8)
}
\arguments{
\item{prcp}{vector of daily precipitation}

\item{months}{vector of months corresponding to daily precipitation}

\item{dry_wet}{threshold precipitation amount for dry/wet states}

\item{wet_extreme_quantile}{threshold quantile for wet/extreme states}
}
\description{
Determine Markov State Thresholds from Precipitation
}

