% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{netcdf.as.sf}
\alias{netcdf.as.sf}
\title{Transform netcdf4 file to sf object}
\usage{
netcdf.as.sf(netcdf_filename, na.rm = T, add.id = F, id.column = "point_id")
}
\description{
Transform netcdf4 file to sf object.
}
