% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{netcdf.extract.points.as.tibble}
\alias{netcdf.extract.points.as.tibble}
\title{Extract specific points from netcdf4, as tibble}
\usage{
netcdf.extract.points.as.tibble(netcdf_filename, points_to_extract)
}
\description{
Extract specific points from a netcdf4 file, as a tibble object.
}
