% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_arimas_monte.R
\name{plot_arimas_monte}
\alias{plot_arimas_monte}
\title{Statistical Distribution Plot of Monte Carlo ARIMA Simulations}
\usage{
plot_arimas_monte(obs, wgen_mc)
}
\arguments{
\item{obs}{numeric vector of historical annual precipitation}

\item{wgen_mc}{output of monte carlo simulation of arimas generated by arimas_monte()}
}
\description{
Statistical Distribution Plot of Monte Carlo ARIMA Simulations
}

