% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gws_power.R
\name{plot_gws_power}
\alias{plot_gws_power}
\title{GWS Power Spectrum Comparison Plot}
\usage{
plot_gws_power(wgen_wave, wgen_mc)
}
\arguments{
\item{wgen_wave}{output of wavelet analysis of historical data generated by wavelet_analysis()}

\item{wgen_mc}{output of monte carlo simulation of arimas generated by arimas_monte()}
}
\description{
GWS Power Spectrum Comparison Plot
}

