% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery.R
\name{recover_data_of_previous_runs}
\alias{recover_data_of_previous_runs}
\title{Recover partial runs}
\usage{
recover_data_of_previous_runs(
  run_output_folder = getwd(),
  run_output_filename = "sim_results.csv",
  avbl_cores = 1
)
}
\arguments{
\item{run_output_folder}{the output_folder which was set to the execution to be recovered}

\item{run_output_filename}{the output_filename which was set to the execution to be recovered}
}
\description{
This feature enables recovery from unexpected situations.
}
