% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save_wgen_to_zip}
\alias{save_wgen_to_zip}
\title{Save weather generator output to zip of csv files}
\usage{
save_wgen_to_zip(x)
}
\arguments{
\item{x}{list of results from gen_month_arma_trend()}
}
\value{
absolute path to zip file
}
\description{
Save weather generator output to zip of csv files
}
\examples{
clim.mon <- get_maurer_mon(lat=42, lon=-72)
x <- gen_month_arma_trend(x.reg.yr=clim.mon, x.loc.mon=clim.mon, n.iter=10, n.year=50, temp.factors=seq(0, 5), prcp.factors=seq(-0.5, 0.5, by=0.5))
save_wgen_to_zip(x)
}

