% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wgen_daily.R
\name{select_knn_day}
\alias{select_knn_day}
\title{Select k-Nearest Neighbor for Given Simulation State}
\usage{
select_knn_day(wday, state, state_prev, prcp_prev, temp_prev, prcp_sd, temp_sd,
  historical)
}
\arguments{
\item{wday}{simulation water-day}

\item{state}{simulated state}

\item{state_prev}{simulated state of previous water-day}

\item{prcp_prev}{simulated precip of previous water-day}

\item{temp_prev}{simulated temp of previous water-day}

\item{prcp_sd}{standard deviation of historical precipitation for current month}

\item{temp_sd}{standard deviation of historical temperature for current month}

\item{historical}{historical climate dataframe}
}
\description{
Select k-Nearest Neighbor for Given Simulation State
}

