% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim.glmwgen}
\alias{sim.glmwgen}
\title{Simulates new weather trajectories in stations}
\usage{
sim.glmwgen(
  model,
  simulation_locations,
  start_date,
  end_date,
  control = glmwgen:::glmwgen_simulation_control()
)
}
\arguments{
\item{model}{A glmwgen model.}

\item{simulation_locations}{a sf object with the points at which weather should be simulated.
If not set, the locations used to fit the model will be used.}

\item{start_date}{a start date in text format (will be converted using as.Date) or a date object.}

\item{end_date}{an end date in text format (will be converted using as.Date) or a date object.}

\item{control}{a glmwgen simulation control list.}
}
\description{
Simulates new weather trajectories.
}
