% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_locs.R
\name{sim.locs.glmwgen}
\alias{sim.locs.glmwgen}
\title{Simulates new weather trajectories in interpolated locations}
\usage{
sim.locs.glmwgen(object, nsim = 1, seed = NULL, start_date = NA,
  end_date = NA, simulation_locations = NULL,
  control = glmwgen:::glmwgen_simulation_control(), verbose = T)
}
\arguments{
\item{object}{A glmwgen model.}

\item{nsim}{number of response vectors to simulate. Defaults to 1.}

\item{seed}{an object specifying if and how the random number generator should be initialized (‘seeded’).

         Either NULL or an integer that will be used in a call to set.seed before simulating the response vectors.
         If set, the value is saved as the "seed" attribute of the returned value. The default, NULL will not change the random generator state.}

\item{start_date}{a start date in text format (will be converted using as.Date) or a date object.}

\item{end_date}{an end date in text format (will be converted using as.Date) or a date object.}

\item{simulation_locations}{a SpatialPoints object with the points at which weather should be simulated.
If not set, the locations used to fit the model will be used.}

\item{control}{a glmwgen simulation control list.}
}
\description{
Simulates new weather trajectories.
}
