% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_annual_arima.R
\name{sim_annual_arima}
\alias{sim_annual_arima}
\title{Simulate Annual Timeseries with ARIMA Model}
\usage{
sim_annual_arima(x, start_water_year = 2000, n_year = 10)
}
\arguments{
\item{x}{vector of observed annual values}

\item{start_water_year}{first simulation water year}

\item{n_year}{number of simulation years}
}
\value{
list containing ARIMA model, observed values, and simulated values as zoo object
}
\description{
Simulate Annual Timeseries with ARIMA Model
}

