% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_daily.R
\name{sim_daily}
\alias{sim_daily}
\title{Run Daily Weather Simulation}
\usage{
sim_daily(historical, n_year, dry_wet_threshold = 0.3,
  wet_extreme_quantile_threshold = 0.8, start_month = 10,
  start_water_year = 2000, include_leap_days = FALSE,
  adjust_annual_precip = TRUE, annual_precip_adjust_limits = c(0.9, 1.1),
  dry_spell_changes = 1, wet_spell_changes = 1, prcp_mean_changes = 1,
  prcp_cv_changes = 1, temp_changes = 0)
}
\arguments{
\item{historical}{historical climate dataframe}

\item{n_year}{number of simulation years}

\item{dry_wet_threshold}{threshold precipitation amount for dry/wet states}

\item{wet_extreme_quantile_threshold}{threshold quantile for wet/extreme states}

\item{start_month}{initial month of the water year}

\item{start_water_year}{initial water year of simulation}

\item{include_leap_days}{include leap days in simulation time series}
}
\value{
a named list containing:
\item{\code{x}}{the historical dataset used to train the simulation}
\item{\code{state_thresholds}}{monthly precipitation thresholds for defining Markov states based on the historical dataset}
\item{\code{transition_matrices}}{monthly transition matrices based on the historical dataset}
\item{\code{state_equilibria}}{monthly state equilibria probabilities}
\item{\code{out}}{a data frame of the simulated daily weather}
}
\description{
Run Daily Weather Simulation
}

