% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_loc.sf}
\alias{sim_loc.sf}
\title{Create simulation locations simple feature (sf)}
\usage{
sim_loc.sf(station_id, longitude, latitude, crs, ...)
}
\description{
Create simulation locations simple feature (sf).
}
