% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_local_result_to_tibble}
\alias{sim_local_result_to_tibble}
\title{Transform simulation result to tibble}
\usage{
sim_local_result_to_tibble(sim_result, nsim_to_extract = 1)
}
\description{
This function transforms the simulation result, a named array, to a tibble.
}
