% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_mc_knn_day.R
\name{sim_mc_knn_day}
\alias{sim_mc_knn_day}
\title{Run Daily Simulation using KNN/Markov Chain}
\usage{
sim_mc_knn_day(n_year, historical, states, transitions, start_month = 10)
}
\arguments{
\item{n_year}{number of simulation years}

\item{historical}{historical climate dataframe}

\item{states}{character list of markov states}

\item{transitions}{list of monthly transition matrices generated with mc_fit()}

\item{start_month}{first month of the water year}
}
\description{
Run Daily Simulation using KNN/Markov Chain
}

