% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate_arimas}
\alias{simulate_arimas}
\title{Run simulations of multiple arima models}
\usage{
simulate_arimas(models, n, components = TRUE)
}
\arguments{
\item{models}{list of arima model objects (e.g. noise and wavelet components)}

\item{n}{number of simulation timesteps}

\item{components}{if TRUE returns list containing both the sum and indidividual components, otherwise returns just sum vector}
}
\description{
Run simulations of multiple arima models
}
\examples{
library(forecast)
models <- list(COMPONENT=auto.arima(sin(1:50)+rnorm(50)+10, max.p=2, max.q=2, max.P=0, max.Q=0, stationary=TRUE),
               NOISE=auto.arima(rnorm(50), max.p=2, max.q=2, max.P=0, max.Q=0, stationary=TRUE))
simulate_arimas(models, n=40)
}

