% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spatial.R
\name{spatial_calibrate}
\alias{spatial_calibrate}
\title{Weather model calibration}
\usage{
spatial_calibrate(
  climate,
  stations,
  seasonal_climate = NULL,
  control = gamwgen:::spatial_fit_control(),
  verbose = F
)
}
\description{
Fits a weather model from historical data.
}
