% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{summarise_seasonal_climate}
\alias{summarise_seasonal_climate}
\title{Generate seasonal covariates}
\usage{
summarise_seasonal_climate(datos_climaticos, umbral_faltantes = 0.2)
}
\description{
Generate seasonal covariates.
}
