% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{SummaryTablePrecipitation}
\alias{SummaryTablePrecipitation}
\title{SummaryTablePrecipitation}
\usage{
SummaryTablePrecipitation(prcp_occ_fit = NULL, prcp_amt_fit = NULL)
}
\description{
Summary precipitation model.
}
