% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trends_mon}
\alias{trends_mon}
\title{Add multiple climate trends to generated monthly climate timeseries}
\usage{
trends_mon(sim.mon.list, temp.factors, prcp.factors)
}
\arguments{
\item{temp.factors}{vector of additive temperature trend factors}

\item{prcp.factors}{vector multiplicative precipitation trend factors}

\item{sim.gen.list}{list of data.frames containing multiple monthly climate timeseries generated by gen_month_arma()}
}
\description{
Add multiple climate trends to generated monthly climate timeseries
}
\examples{
sim.gen <- gen_month_arma(x.reg.yr=clim.reg.yr, x.loc.mon=clim.loc.mon, n.iter=50, n.year=nrow(clim.reg.yr))
trends_mon(sim.mon.list=sim.gen$sim.mon, temp.factors=seq(0, 2, by=0.5), prcp.factors=seq(0.8, 1.2, by=0.1))
}

