% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/waterday.R
\name{waterday}
\alias{waterday}
\title{Get julian day component by water year of a date-time (aka water day)}
\usage{
waterday(x, start_month = 10)
}
\arguments{
\item{x}{a vector of date-times}

\item{start_month}{initial month of the water year (default is 10 for October)}
}
\description{
Get julian day component by water year of a date-time (aka water day)
}
\examples{
waterday(lubridate::ymd("2000-10-01"), start_month=10)
waterday(lubridate::ymd(c("2000-01-01", "2000-09-30", "2000-10-01", "2000-12-31")), start_month=10)
}

