% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{WAVELET_ANALYSIS}
\alias{WAVELET_ANALYSIS}
\title{k-Nearest Neighbors of monthly climate}
\usage{
WAVELET_ANALYSIS(CLIMATE_VARIABLE, siglvl, background_noise, plot_flag)
}
\arguments{
\item{CLIMATE_VARIABLE}{annual timeseries of observed climate variable (e.g. precipitation)}

\item{siglvl}{significance level}

\item{background_noise}{type of background noise (white, red)}

\item{plot_flag}{boolean flag to show plots}
}
\description{
k-Nearest Neighbors of monthly climate
}
\examples{
WAVELET_ANALYSIS(CLIMATE_VARIABLE,siglvl,background_noise,plot_flag)
}

