% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{wavelet_annual}
\alias{wavelet_annual}
\title{Wavelet analysis of annual timeseries}
\usage{
wavelet_annual(x, sig = 0.9, noise.type = "white", plot.flag = TRUE)
}
\arguments{
\item{x}{annual values of climate timeseries (e.g. precipitation) as numeric array}

\item{sig}{significance level}

\item{noise.type}{type of background noise ("white" or "red")}

\item{plot.flag}{boolean flag to show plots}
}
\description{
Wavelet analysis of annual timeseries
}
\examples{
wavelet_annual(x=c(354.2, 234.4, 343), sig=0.90, noise.type="white", plot.flag=TRUE)
}

