% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{wavelet_decomposition}
\alias{wavelet_decomposition}
\title{Wavelet decomposition}
\usage{
wavelet_decomposition(x, n.periods, sig.periods, n.comp.periods,
  plot_flag = TRUE)
}
\arguments{
\item{x}{annual timeseries of observed climate variable (e.g. precipitation)}

\item{n.periods}{number of final periods as integer}

\item{sig.periods}{integer array of significance periods}

\item{n.comp.periods}{number of comparison periods}

\item{plot_flag}{boolean flag to show plots}
}
\description{
Wavelet decomposition
}
\examples{
wavelet_decomposition(x=x, n.periods=1, sig.periods=seq(8,15), n.comp.periods=8, plot_flag=TRUE)
}

