% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{wavelet_morlet}
\alias{wavelet_morlet}
\title{Apply Morlet wavelet function to Fourier frequencies}
\usage{
wavelet_morlet(k, s)
}
\arguments{
\item{k}{numeric array of Fourier frequencies at which to compute wavelet}

\item{s}{wavelet scale parameter}
}
\value{
List containing

\itemize{
  \item \code{daughter}: a numeric vector of the wavelet function
  \item \code{fourier_factor}: ratio of Fourier period to scale
  \item \code{coi}: cone of influence
  \item \code{dofmin}: degrees of freedom for each point in wavelet power
}
}
\description{
Apply Morlet wavelet function to Fourier frequencies
}
\examples{
waveletf(k,s)
}
\references{
http://paos.colorado.edu/research/wavelets/software.html
}

