% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{waveletf}
\alias{waveletf}
\title{Wavelet function that returns transform}
\usage{
waveletf(k, s)
}
\arguments{
\item{k}{wavelet number array}

\item{s}{scale parameter}
}
\description{
Wavelet function that returns transform
}
\examples{
waveletf(k,s)
}

