% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{waveletf2}
\alias{waveletf2}
\title{Wavelet function that returns fourier_factor, cone of influence and degrees of freedom}
\usage{
waveletf2(k, s)
}
\arguments{
\item{k}{wavelet number array}

\item{s}{scale parameter}
}
\description{
Wavelet function that returns fourier_factor, cone of influence and degrees of freedom
}
\examples{
waveletf2(k,s)
}
\references{
http://paos.colorado.edu/research/wavelets/software.html
}

