\name{weathergen-package}
\alias{weathergen-package}
\alias{weathergen}
\docType{package}
\title{
Weather Generator
}
\description{
Daily weather generator for evaluating system vulnerability to climate change.
}
\details{
\tabular{ll}{
Package: \tab weathergen\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-07-23\cr
License: \tab What license is it under?\cr
}
Use this package to generate synthetic time series of precipitation and air temperature based on observed datasets.
}
\author{
Jeffrey D. Walker <jeff@walkerjeff.com>
Scott Steinschneider
Maintainer: Jeffrey D. Walker <jeff@walkerjeff.com>
}
\references{
Steinschneider, S., and C. Brown (2013), A semiparametric multivariate, multi-site weather generator with low-frequency variability for use in climate risk assessments, Water Resour. Res., 49, 7205-7220, doi: 10.1002/wrcr.20528.
}
\keyword{ package multivariate }
\seealso{
}
\examples{
DAILY_WEATHER_GENERATOR()
}
