% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wgen_daily.R
\name{wgen_daily}
\alias{wgen_daily}
\title{Run Daily Weather Simulation}
\usage{
wgen_daily(obs_day, n_year, start_month = 10, start_water_year = 2000,
  include_leap_days = FALSE, n_knn_annual = 100, dry_wet_threshold = 0.3,
  wet_extreme_quantile_threshold = 0.8, adjust_annual_precip = TRUE,
  annual_precip_adjust_limits = c(0.9, 1.1), dry_spell_changes = 1,
  wet_spell_changes = 1, prcp_mean_changes = 1, prcp_cv_changes = 1,
  temp_mean_changes = 0)
}
\arguments{
\item{obs_day}{daily historical observation dataset as \code{zoo} object with variables PRCP, TEMP, TMIN, TMAX, WIND}

\item{n_year}{number of simulation years}

\item{start_month}{initial month of the water year}

\item{start_water_year}{initial water year of simulation}

\item{include_leap_days}{include leap days in simulation time series}

\item{n_knn_annual}{number of years used in knn sampling algorithm}

\item{dry_wet_threshold}{threshold precipitation amount for dry/wet states}

\item{wet_extreme_quantile_threshold}{threshold quantile for wet/extreme states}

\item{adjust_annual_precip}{flag to adjust simulated daily precip to match simulated annual precip}

\item{annual_precip_adjust_limits}{range of maximum annual precip adjustment factors}

\item{dry_spell_changes}{adjustment factor(s) for dry spell durations (single value, or vector of length 12 for monthly)}

\item{wet_spell_changes}{adjustment factor(s) for wet spell durations (single value, or vector of length 12 for monthly)}

\item{prcp_mean_changes}{adjustment factor(s) for mean precip (single value, or vector of length 12 for monthly)}

\item{prcp_cv_changes}{adjustment factor(s) for precip CV (single value, or vector of length 12 for monthly)}

\item{temp_mean_changes}{adjustment factor(s) for mean temp (single value, or vector of length 12 for monthly)}
}
\value{
a named list containing:
\item{\code{obs}}{the historical observation dataset used to train the simulation}
\item{\code{state_thresholds}}{monthly precipitation thresholds for defining Markov states based on the historical dataset}
\item{\code{transition_matrices}}{monthly transition matrices based on the historical dataset}
\item{\code{state_equilibria}}{monthly state equilibria probabilities}
\item{\code{sim}}{a data frame of the simulated daily weather}
}
\description{
Run Daily Weather Simulation
}

