% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wgen_daily_from_annual.R
\name{wgen_daily_from_annual}
\alias{wgen_daily_from_annual}
\title{Simulate Daily Weather from Simulated Annual Precipitation}
\usage{
wgen_daily_from_annual(prcp_yr, obs_da, obs_prcp_yr, start_month = 10,
  include_leap_days = FALSE, n_annual_knn = 100)
}
\arguments{
\item{obs_da}{data frame of daily observed weather}

\item{start_month}{first month of the water year}

\item{include_leap_days}{boolean flag to include leap days (default=FALSE)}

\item{n_annual_knn}{number of years for annual KNN sampling}

\item{sim_yr_prcp}{vector of simulated annual precipitation}

\item{sim_yr_years}{vector of years corresponding to sim_yr_prcp}

\item{obs_yr_prcp}{vector of observed annual precipitation}

\item{obs_yr_years}{vector of years corresponding to obs_yr_prcp}

\item{obs_da_years}{vector of years corresponding to obs_da_prcp}
}
\value{
vector of sampled years of length n
}
\description{
Simulate Daily Weather from Simulated Annual Precipitation
}

