% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wgen_historical_day.R
\name{wgen_historical_day}
\alias{wgen_historical_day}
\title{Run Daily Weather Generator using Historical Data}
\usage{
wgen_historical_day(historical, n_year, start_month = 10,
  start_water_year = 2000, include_leap_days = FALSE)
}
\arguments{
\item{historical}{historical climate dataframe}

\item{n_year}{number of simulation years}

\item{start_month}{initial month of the water year}

\item{start_water_year}{initial water year of simulation}

\item{include_leap_days}{include leap days in simulation time series}
}
\description{
Run Daily Weather Generator using Historical Data
}

