% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjust_daily_to_annual.R
\name{adjust_daily_to_annual}
\alias{adjust_daily_to_annual}
\title{Adjust Daily Values to Match Annual Sums}
\usage{
adjust_daily_to_annual(x_day, years_day, x_yr, years_yr, min_ratio = 0.9,
  max_ratio = 1.1)
}
\arguments{
\item{x_day}{vector of daily values}

\item{years_day}{vector of years corresponding to x_day}

\item{x_yr}{vector of annual sums}

\item{years_yr}{vector of years corresponding to x_yr}

\item{min_ratio}{minimum adjustment ratio}

\item{max_ratio}{maximum adjustment ratio}
}
\description{
Adjust Daily Values to Match Annual Sums
}

