% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_maurer_mon}
\alias{get_maurer_mon}
\title{Retrieve monthly timeseries for given location from maurer dataset}
\usage{
get_maurer_mon(lat, lon)
}
\arguments{
\item{lat}{latitude in decimal degrees North}

\item{lon}{longitude in decimal degrees East (negative for West)}
}
\description{
Retrieve monthly timeseries for given location from maurer dataset
}
\examples{
get_maurer_mon(42, -71)
}

