% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{glmwgen_fit_control}
\alias{glmwgen_fit_control}
\title{Weather model fit configuration}
\usage{
glmwgen_fit_control(prcp_occurrence_threshold = 0.1,
  use_external_seasonal_climate = T, climate_missing_threshold = 0.2,
  use_covariates = F, avbl_cores = 2)
}
\description{
Provides fine control of different parameters that will be used to fit a weather model.
}
