% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{glmwgen_fit_control}
\alias{glmwgen_fit_control}
\title{Weather model fit configuration}
\usage{
glmwgen_fit_control(prcp_occurrence_threshold = 0.1,
  use_seasonal_covariates = T, use_linear_term = F,
  seasonal_covariates = c("tx", "tn", "prcp"), save_residuals = F,
  cov_mode = "pairwise.complete")
}
\description{
Provides fine control of different parameters that will be used to fit a weather model.
}

