% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ctrl.R
\name{glmwgen_simulation_control}
\alias{glmwgen_simulation_control}
\title{Simulations control configuration}
\usage{
glmwgen_simulation_control(Rt = NULL, multicombine = F,
  cache_size = 1)
}
\description{
Provides fine control of different parameters that will be used to create new weather series.
}
