% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{glmwgen_simulation_control}
\alias{glmwgen_simulation_control}
\title{Simulations control configuration}
\usage{
glmwgen_simulation_control(
  nsim = 1,
  seed = NULL,
  avbl_cores = 2,
  bbox_offset = 1e+05,
  sim_loc_as_grid = T,
  use_spatially_correlated_noise = T,
  use_temporary_files_to_save_ram = T,
  remove_temp_files_used_to_save_ram = T,
  manage_parallelization_externally = F
)
}
\arguments{
\item{nsim}{number of response vectors to simulate. Defaults to 1.}

\item{seed}{an object specifying if and how the random number generator should be initialized (‘seeded’).#'
Either NULL or an integer that will be used in a call to set.seed before simulating the response vectors.
If set, the value is saved as the "seed" attribute of the returned value. The default, NULL will not change the random generator state.}

\item{avbl_cores}{...}

\item{coord_ref_system}{...}
}
\description{
Provides fine control of different parameters that will be used to create new weather series.
}
