% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ctrl.R
\name{glmwgen_simulation_control}
\alias{glmwgen_simulation_control}
\title{Simulations control configuration}
\usage{

  glmwgen_simulation_control(seasonal_temps_covariates_getter = get_temperatures_seasonal_covariate,
  seasonal_prcp_covariates_getter = get_rainfall_seasonal_covariate,
  random_fields_method = "rn", Rt = NULL, multicombine = F,
  always_krig_coefficients = T, interpolation_method = c("idw",
  "kriging"), cache_size = 1,
  minimum_temperatures_difference_threshold = 0.1)
}
\description{
Provides fine control of different parameters that will be used to create new weather series.
}
