% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knn.R
\name{knn_daily}
\alias{knn_daily}
\title{k-Nearest Neighbors of Daily Data}
\usage{
knn_daily(cur_sim_PRCP, cur_sim_TEMP, PRCP_TODAY, TEMP_TODAY, PRCP_TOMORROW,
  TEMP_TOMORROW, TMAX_TOMORROW, TMIN_TOMORROW, DATE_TOMORROW, k,
  sd_monthly_PRCP, sd_monthly_TEMP, mean_monthly_PRCP, mean_monthly_TEMP, k1,
  count)
}
\arguments{
\item{x}{annual values of climate timeseries (e.g. precipitation) as numeric array}
}
\description{
k-Nearest Neighbors of Daily Data
}
\examples{
knn_daily(x=runif(30)*100+200, sig=0.90, noise.type="white", plot.flag=TRUE)
}

