% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_local.R
\name{local_calibrate}
\alias{local_calibrate}
\title{Weather model calibration}
\usage{
local_calibrate(
  climate,
  stations,
  seasonal_climate = NULL,
  control = glmwgen:::local_fit_control(),
  verbose = F
)
}
\description{
Fits a weather model from historical data.
}
