% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_local.R
\name{local_simulation}
\alias{local_simulation}
\title{Simulates new weather trajectories in stations}
\usage{
local_simulation(
  model,
  simulation_locations,
  start_date,
  end_date,
  control = gamwgen:::local_simulation_control(),
  output_filename = "sim_results.rds",
  seasonal_covariates = NULL,
  verbose = F
)
}
\arguments{
\item{model}{A gamwgen model.}

\item{simulation_locations}{a sf object with the points at which weather should be simulated.
If not set, the locations used to fit the model will be used.}

\item{start_date}{a start date in text format (will be converted using as.Date) or a date object.}

\item{end_date}{an end date in text format (will be converted using as.Date) or a date object.}

\item{control}{a gamwgen simulation control list.}
}
\description{
Simulates new weather trajectories.
}
