% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{netcdf.as.tibble}
\alias{netcdf.as.tibble}
\title{Transform netcdf4 file to tibble}
\usage{
netcdf.as.tibble(
  netcdf_filename,
  points_to_extract,
  points_id_column = "point_id"
)
}
\description{
Transform netcdf4 file to tibble.
}
