% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_annual_arima.R
\name{sim_annual_arima}
\alias{sim_annual_arima}
\title{Simulate Annual Timeseries with ARIMA Model}
\usage{
sim_annual_arima(x, start_year = 2000, n_year = 10)
}
\arguments{
\item{x}{vector of observed annual values}

\item{start_year}{first simulation year}

\item{n_year}{number of simulation years}
}
\value{
a named list containing:
\item{\code{model}}{the fitted arima model returned by \code{forecast::auto.arima()}}
\item{\code{x}}{the historical annual precipitation as a numeric vector}
\item{\code{out}}{the simulated annual precipitation of length \code{n_year} as a \code{zoo} object}
}
\description{
Fit an ARIMA model to a given vector of annual precipitation \code{x}, and run a simulation of length \code{n_year} beginning with year \code{start_year}.
}
\examples{
sim_annual_arima(rnorm(100))
}

