% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_daily_from_annual.R
\name{sim_daily_from_annual}
\alias{sim_daily_from_annual}
\title{Simulate Daily Weather from Simulated Annual Precipitation}
\usage{
sim_daily_from_annual(prcp_yr, obs_da, obs_prcp_yr, start_month = 10,
  include_leap_days = FALSE, n_knn_annual = 100)
}
\arguments{
\item{prcp_yr}{vector of target annual precipitation}

\item{obs_da}{data frame of daily observed weather}

\item{obs_prcp_yr}{vector of observed annual precipitation}

\item{start_month}{first month of the water year}

\item{include_leap_days}{boolean flag to include leap days (default=FALSE)}

\item{n_knn_annual}{number of years for annual KNN sampling}
}
\value{
vector of sampled years of length n
}
\description{
Simulate Daily Weather from Simulated Annual Precipitation
}

