% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_mc_knn_day.R
\name{sim_mc_knn_day}
\alias{sim_mc_knn_day}
\title{Run Daily Simulation using KNN/Markov Chain}
\usage{
sim_mc_knn_day(x, n_year, states, transitions, start_month = 10,
  start_water_year = 2000, include_leap_days = FALSE)
}
\arguments{
\item{x}{historical daily climate dataset}

\item{n_year}{number of simulation years}

\item{states}{character list of markov states}

\item{transitions}{list of monthly transition matrices generated with mc_fit()}

\item{start_month}{first month of the water year}

\item{start_water_year}{initial water year of simulation}

\item{include_leap_days}{include leap days in simulation time series}
}
\description{
Run Daily Simulation using KNN/Markov Chain
}

