% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate_arima}
\alias{simulate_arima}
\title{Run simulation of an arima model}
\usage{
simulate_arima(model, n)
}
\arguments{
\item{model}{arima model object}

\item{n}{number of simulation timesteps}
}
\description{
Run simulation of an arima model
}
\examples{
library(forecast)
model <- auto.arima(sin(1:50)+rnorm(50)+10, max.p=2, max.q=2, max.P=0, max.Q=0, stationary=TRUE)
simulate_arima(model, n=40)
}

