% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{trends_mon}
\alias{trends_mon}
\title{Add multiple climate trends to generated monthly climate timeseries}
\usage{
trends_mon(sim.mon.list, temp.factors, prcp.factors)
}
\arguments{
\item{sim.gen.list}{list of data.frames containing multiple monthly climate timeseries generated by gen_month_arma()}

\item{temp.factor}{additive temperature trend factor}

\item{prcp.factor}{multiplicative precipitation trend factor}
}
\description{
Add multiple climate trends to generated monthly climate timeseries
}
\examples{
sim.gen <- gen_month_arma(x.reg.yr=clim.reg.yr, x.loc.mon=clim.loc.mon, n.iter=50, n.year=nrow(clim.reg.yr))
trends_mon(sim.mon.list=sim.gen$sim.mon, temp.factor=2, prcp.factor=1.5)
}

