% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterday_range.R
\name{waterday_range}
\alias{waterday_range}
\title{Get a range of water days around some center value}
\usage{
waterday_range(x, n = 7)
}
\arguments{
\item{x}{central water day}

\item{n}{length of range}
}
\description{
Get a range of water days around some center value
}
\examples{
waterday_range(x=1, n=7)
waterday_range(x=10, n=7)
waterday_range(x=363, n=7)
}

