% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wavelet_analysis.R
\name{wavelet_analysis}
\alias{wavelet_analysis}
\title{Wavelet Analysis of Annual Timeseries}
\usage{
wavelet_analysis(x, sig.level = 0.9, noise.type = c("white", "red"))
}
\arguments{
\item{x}{annual values of climate timeseries (e.g. precipitation) as numeric array}

\item{sig.level}{significance level}

\item{noise.type}{type of background noise ("white" or "red")}
}
\description{
Wavelet Analysis of Annual Timeseries
}
\examples{
wavelet_analysis(x=runif(30)*100+200, sig=0.90, noise.type="white", plot.flag=TRUE)
}

