% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wavelet_analysis}
\alias{wavelet_analysis}
\title{Wavelet analysis of annual timeseries}
\usage{
wavelet_analysis(x, sig.level = 0.9, noise.type = c("white", "red"))
}
\arguments{
\item{x}{annual values of climate timeseries (e.g. precipitation) as numeric array}

\item{sig.level}{significance level}

\item{noise.type}{type of background noise ("white" or "red")}
}
\description{
Wavelet analysis of annual timeseries
}
\examples{
wavelet_analysis(x=runif(30)*100+200, sig=0.90, noise.type="white", plot.flag=TRUE)
}

