% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wavelet_components}
\alias{wavelet_components}
\title{Reconstruct wavelet components}
\usage{
wavelet_components(x, wt, n.periods, sig.periods, n.comp.periods)
}
\arguments{
\item{x}{annual timeseries of observed climate variable as numeric vector}

\item{wt}{wavelet transform generated by wavelet_analysis()}

\item{n.periods}{number of final composite periods}

\item{sig.periods}{integer array of significance periods}

\item{n.comp.periods}{number of significance periods for each composite period}

\item{plot_flag}{boolean flag to show plots}
}
\description{
Reconstruct wavelet components
}
\examples{
wavelet_components(x=x, n.periods=1, sig.periods=seq(8,15), n.comp.periods=8)
}

