% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{WAVELET_DECOMPOSITION}
\alias{WAVELET_DECOMPOSITION}
\title{Wavelet decomposition}
\usage{
WAVELET_DECOMPOSITION(CLIMATE_VARIABLE, NUM_FINAL_PERIODS, ALL_SIG_PERIODS,
  NUM_PERIODS_ALL_COMPS, plot_flag)
}
\arguments{
\item{CLIMATE_VARIABLE}{annual timeseries of observed climate variable (e.g. precipitation)}

\item{NUM_FINAL_PERIODS}{number of final periods as integer}

\item{ALL_SIG_PERIODS}{integer array of significance periods}

\item{NUM_PERIODS_ALL_COMPS}{number of comparison periods}

\item{plot_flag}{boolean flag to show plots}
}
\description{
Wavelet decomposition
}
\examples{
WAVELET_DECOMPOSITION(CLIMATE_VARIABLE,NUM_FINAL_PERIODS,ALL_SIG_PERIODS,NUM_PERIODS_ALL_COMPS,plot_flag)
}

