% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wgen_daily.R
\name{wgen_historical_day}
\alias{wgen_historical_day}
\title{Run Daily Weather Generator using Historical Data}
\usage{
wgen_historical_day(historical, n_year, start_mon = 10)
}
\arguments{
\item{historical}{historical climate dataframe}

\item{n_year}{number of simulation years}

\item{start_mon}{initial month of the water year}
}
\description{
Run Daily Weather Generator using Historical Data
}

