% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{wyear}
\alias{wyear}
\title{Wavelet analysis of annual timeseries}
\usage{
wyear(x, start.month = 10)
}
\arguments{
\item{x}{an array of dates in formats accepted by lubridate}

\item{start.month}{initial month of the water year (default is 10 for October)}
}
\description{
Wavelet analysis of annual timeseries
}
\examples{
library(lurbidate)
wyear(ymd(c("2000-01-01", "2000-09-30", "2000-10-01", "2000-12-31")), start.month=10)
}

